sql-to-medallion/requirements.txt
# Core dependencies
pyspark==3.4.1
delta-spark==2.4.0
pyarrow==12.0.1
pandas==2.0.3
numpy==1.25.2

# Azure dependencies
azure-storage-file-datalake==12.12.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0

# SQL Server connectivity
pymssql==2.2.8
jaydebeapi==1.2.3

# OpenAI integration
openai==0.28.1
azure-ai-ml==1.11.1
azure-mgmt-cognitiveservices==13.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
loguru==0.7.0
jinja2==3.1.2

# Development dependencies
pytest==7.4.2
pytest-mock==3.11.1
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Optional for SQL Server JDBC connection (ensure driver is available in Databricks)
# pyodbc # Typically not needed if using Spark JDBC directly

# For development/notebooks
jupyterlab 